enhanced type choosing for League groups
Elite 4, Champions, and Gyms now should not share types with each other, unless there are too many Gyms to assign each an unused type. In this case, the first eight (nine, actually) gyms are guaranteed to not share types with the other groups, and later gyms will not share types with other gyms.