(Gen 6) fix XY failing ubiquitous move test
The issue had to do with Romhandler.getMoveTutorCompatibility()
crashing with XY.
This is weird, since XY *does* have move tutors,
and these *should* be supported. Assuming though
that the ZX team ran into troubles reading them,
and decided to mark XY off as hasMoveTutors()=false.

It works fine enough, as long as you know not to
get MT data for games that doesn't support it.