separate pkromio from pkrandom
Using this structure, NO code in pkromio should reference pkrandom.

This way the RomHandler interface, and all associated data,
can be used for other projects without being bogged down
by Randomizer code.
Plus, it's just a neater way to structure the code.

In the long term, this might lead to pkromio being broken
out into its own repo, but this remains to be seen.

This being the initial separation, there likely are some stragglers.
PaletteDescriptionTool is one known issue to be dealt with.